requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
python-decouple==3.5
pydantic>=1.8.0,<2.0.0
requests==2.26.0
imread-from-url==0.1.3
matplotlib==3.5.1
numpy==1.23.4
onnxruntime-gpu==1.13.1
opencv_python==4.5.4.60
# pandas==1.3.3
# Pillow==8.4.0
# PyYAML==6.0
# scipy==1.7.1
# seaborn==0.11.2
# thop==0.0.31.post2005241907
# tqdm==4.62.3